refactor(components/molecule/tabs): api tab
